requirements.txt
openai-agents>=0.2.0
pydantic>=2.7
fastapi>=0.111
uvicorn>=0.30
python-dotenv>=1.0
httpx>=0.27
chromadb>=0.4.22
numpy>=1.24.0
